refactor(mull-ui): move home page subpages into home directory
allowed me to put all repeated styles into home.discover.scss, reducing repeated code

Refs #65
