+++ apps/mull-ui/src/app/pages/home/discover/discover.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { IEvent } from '@mull/types';
-import { EventCard } from '../../components';
+import { EventCard } from '../../../components';
 import { gql, useQuery } from '@apollo/client';
 
-import './discover.scss';
+import '../home-discover.scss';
 
 export const DiscoverPage = ({ history }) => {
   const GET_DISCOVER_EVENTS = gql`
