--- apps/mull-ui/src/app/pages/upcoming/upcoming.scss
@@ -1,9 +1,0 @@
-.discover-page-tabs-container {
-  margin-top: 1.5rem;
-  .event-card-container {
-    margin-bottom: 2rem;
-    &:last-child {
-      margin-bottom: 1rem;
-    }
-  }
-}
